SVM and refactoring
