+++ naive_bayes/nb_author_id.py
@@ -11,31 +11,18 @@
 """
     
 import sys
-from time import time
 
 from sklearn.naive_bayes import GaussianNB, MultinomialNB, BernoulliNB
-from sklearn.metrics import accuracy_score
 
 sys.path.append("../tools/")
 from email_preprocess import preprocess
+from mylib import fit_and_predict
 
 
 ### features_train and features_test are the features for the training
 ### and testing datasets, respectively
 ### labels_train and labels_test are the corresponding item labels
 features_train, features_test, labels_train, labels_test = preprocess()
-
-
-def fit_and_predict(clf, features_train, features_test, labels_train, labels_test):
-    t0 = time()
-    clf.fit(features_train, labels_train)
-    t1 = time()
-    pred = clf.predict(features_test)
-    t2 = time()
-    accuracy = accuracy_score(labels_test, pred)
-    t3 = time()
-    print("{:05.3f} : {:05.3f} : {:05.3f} | {:05.3f}".format(t1-t0, t2-t1, t3-t2, t3-t0))
-    return accuracy
 
 print("{:5} : {:5} : {:5} | {:5}".format("fit", "pred", "acc", "total"))
